full user crud

+++ resources/companys.py
@@ -25,17 +25,26 @@ def create():
 @companys.route('/<id>', methods=['Delete'])
 @login_required
 def delete(id):
-    delete_query = models.Company.delete().where(models.Company.id == id)
-    delete_query.execute()
+    #if current user is master user, allow delete, otherwise don't
+    if (current_user.master):
+        delete_query = models.Company.delete().where(models.Company.id == id)
+        delete_query.execute()
+        return jsonify(data={}, message=f"successfully deleted company with id of {id}", status=200),200
+    else: 
+        return jsonify(data={}, message="you don't have the access rights to do that", status=200), 200
 
-    return jsonify(data={}, message=f"successfully deleted company with id of {id}", status=200),200
+
 
 #Update route, need to get auth working then only master user can do this.
 @companys.route('/<id>', methods=['PUT'])
 @login_required
 def update(id):
+    #allow user to update company only if master user
     payload = request.get_json()
-    update_query = models.Company.update(**payload).where(models.Company.id == id)
-    updated_company = models.Company.get_by_id(id)
-    
-    return jsonify(data={}, message=f'succesfully update company {updated_company.name}', status=200),200
+    if current_user.master:
+        update_query = models.Company.update(**payload).where(models.Company.id == id)
+        update_query.execute()
+        updated_company = models.Company.get_by_id(id)
+        return jsonify(data={}, message=f'succesfully update company {updated_company.name}', status=200),200
+    else:
+        return jsonify(data={}, message="you don't have the access rights to do that", status=200), 200
